backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
CacheControl==0.14.4
cachetools==6.2.4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cohere==5.20.1
cryptography==46.0.3
distro==1.9.0
fastapi==0.115.0
fastavro==1.12.1
filelock==3.20.1
firebase_admin==7.1.0
fsspec==2025.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-firestore==2.22.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-genai==1.56.0
google-generativeai==0.8.6
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.2.3
hyperframe==6.1.0
idna==3.11
msgpack==1.1.2
numpy==2.3.4
packaging==25.0
pillow==10.4.0
proto-plus==1.27.0
protobuf==5.29.5
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
pyparsing==3.2.5
python-dotenv==1.2.1
python-multipart==0.0.9
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.38.6
tenacity==9.1.2
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.1
types-requests==2.32.4.20250913
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
